requirements.txt
annotated-doc
annotated-types
anyio
certifi
cffi
charset-normalizer
click
cryptography
fastapi
filelock
h11
idna
joblib
numpy
pandas
pycparser
pydantic
pydantic_core
python-dateutil
python-whois
requests
requests-file
scikit-learn
scipy
six
starlette
threadpoolctl
tldextract
tranco
typing-inspection
typing_extensions
tzdata
urllib3
uvicorn
xgboost
